+++ src/ssd1306/SSD1306_OLED.cpp
@@ -19,8 +19,6 @@ SSD1306  :: SSD1306(int16_t oledwidth, int16_t oledheight) :SSD1306_graphics(ole
 	_OLED_HEIGHT = oledheight;
 	_OLED_WIDTH = oledwidth;
 	_OLED_PAGE_NUM = (_OLED_HEIGHT/8); 
-	_bufferWidth = _OLED_WIDTH;
-	_bufferHeight = _OLED_HEIGHT;
 }
 
 /*!
@@ -350,8 +348,8 @@ void SSD1306::I2C_Write_Byte(uint8_t value, uint8_t cmd)
 */
 void SSD1306::OLEDupdate()
 {
-	uint8_t x = 0; uint8_t y = 0; uint8_t w = this->_bufferWidth; uint8_t h = this->_bufferHeight;
-	OLEDBuffer( x,  y,  w,  h, (uint8_t*) this->_OLEDbuffer);
+	uint8_t x = 0; uint8_t y = 0; uint8_t w = this->_OLED_WIDTH; uint8_t h = this->_OLED_HEIGHT; 
+	OLEDBuffer( x,  y,  w,  h, this->_OLEDbuffer);
 }
 
 /*!
@@ -359,7 +357,7 @@ void SSD1306::OLEDupdate()
 */
 void SSD1306::OLEDclearBuffer()
 {
-	memset( this->_OLEDbuffer, 0x00, (this->_bufferWidth * (this->_bufferHeight /8))  );
+	memset( this->_OLEDbuffer, 0x00, (this->_OLED_WIDTH * (this->_OLED_HEIGHT /8))  );
 }
 
 /*!
@@ -413,11 +411,12 @@ void SSD1306::OLEDBuffer(int16_t x, int16_t y, uint8_t w, uint8_t h, uint8_t* da
 void SSD1306::drawPixel(int16_t x, int16_t y, uint8_t color)
 {
 
-	if ((x < 0) || (x >= this->_bufferWidth) || (y < 0) || (y >= this->_bufferHeight)) {
+	if ((x < 0) || (x >= this->_width) || (y < 0) || (y >= this->_height)) {
 	return;
 	}
 	int16_t temp;
-	switch (_rotation) {
+	uint8_t rotation = getRotation();
+	switch (rotation) {
 	case 1:
 		temp = x;
 		x = WIDTH - 1 - y;
@@ -433,7 +432,7 @@ void SSD1306::drawPixel(int16_t x, int16_t y, uint8_t color)
 		y = HEIGHT - 1 - temp;
 	break;
 	}
-		uint16_t tc = (_bufferWidth * (y /8)) + x;
+		uint16_t tc = (_OLED_WIDTH * (y /8)) + x;
 		switch (color)
 		{
 			case WHITE:  this->_OLEDbuffer[tc] |= (1 << (y & 7)); break;