+++ extra/doc/CHANGELOG.md
@@ -9,3 +9,8 @@
 		Timeout, retry attempts amd retry delay.
 	* Wrapped the buffer pointer in a function.
 	* Added I2C test example file. 
+* Version 1.0.2 Feb 24
+	* minor update
+	* added rotate enum.
+	* corrected number of characters parameter for sixteen segment fotn 15 to 14.
+	* corrected 'drawPixel' boundary scan for 90 and 270 degree rotation.